feat(nx): use platform certificate trust store (#31999)
<!-- Please make sure you have read the submission guidelines before
posting an PR -->
<!--
https://github.com/nrwl/nx/blob/master/CONTRIBUTING.md#-submitting-a-pr
-->

<!-- Please make sure that your commit message follows our format -->
<!-- Example: `fix(nx): must begin with lowercase` -->

<!-- If this is a particularly complex change or feature addition, you
can request a dedicated Nx release for this pull request branch. Mention
someone from the Nx team or the `@nrwl/nx-pipelines-reviewers` and they
will confirm if the PR warrants its own release for testing purposes,
and generate it for you if appropriate. -->

## Current Behavior

Certificates in the local trust store are ignored, because reqwest
(rustls) is using webpki-roots.

## Expected Behavior

Certificates from the local trust store are trusted.

This PR solves this the same way as turborepo - by turning on the
`rustls-tls-native-roots` feature in reqwest.

(cherry picked from commit 9248dcd9a944ec09fffdb4267726de108e7160e0)