fix(module-federation): do not rely on virtualRuntimeEntry #31831 (#32387)
## Current Behavior
Previously, we needed to rely on `virtualRuntimeEntry` to allow builds
between caching to operate successfully.

## Expected Behavior
In more recent version of `@module-federation/enhanced` this is no
longer the case.
By setting this option, it is now inflating the size of the bundles
produced by Webpack and Rspack
As it is no longer a strict requirement, remove it. 

Users can still set it manually if they run into issues

## Related Issue(s)

Fixes #31831
Fixes #32404