fix(core): fail fast when running commands in parallel (#32386)
## Current Behavior

When running commands in parallel with `nx:run-commands`, if a command
fails, others are not stopped, and the task continues until all commands
finish. This means that if a long-running command is running, it won't
be stopped when other commands have already failed.

## Expected Behavior

`nx:run-commands` should fail fast. In a parallel execution, the rest of
the running commands should be terminated if a command fails.

## Related Issue(s)

Fixes #28477

(cherry picked from commit d23cb37d7ea7935bd91e42e0d2920ef0bc1a3e20)