fix(misc): check for `packages` existence to detect pnpm workspaces setup (#32474)
## Current Behavior

When detecting whether a workspace is set up with pnpm workspaces, we
only check for the existence of the `pnpm-workspace.yaml` file. While
most of the time that's enough, it's not entirely correct because that
file can exist without a `packages` entry and solely contain other
settings.

## Expected Behavior

When detecting whether a workspace is set up with pnpm workspaces, we
should check for the existence of the `pnpm-workspace.yaml` file and the
`packages` entry.

(cherry picked from commit 0494d09e8e39ecc892864b610db3706ecb8e2ddc)