docs(misc): fix issues with base href, rewrites, and connecting next.js and astro docs (#32484)
This PR fixes issues with connecting new Astro docs with the `/docs`
rewrite rule on Next.js side.

https://www.loom.com/share/150deedd25a24259b0aa8cb141cec1d3

The redirects now go through Netlify since it's a lot more flexible, and
we need to do that later when porting Next.js to Netlify, etc.

## Not done / follow-up:
Other pages and UI components that use `<Link>` to a doc page need to
check against `NEXT_PUBLIC_ASTRO_URL` and make sure to use the updated
`/docs` path for that. I only handled the homepage for now just to show
it working.