feat(core): replace the pagination from the tui tasks list with scrolling (#32560)
## Current Behavior

The TUI tasks list uses pagination when it has many tasks that don't fit
in the viewport.

## Expected Behavior

The TUI tasks list should be scrollable and not use pagination.

Additionally, this PR includes a fix so that when a task is selected in
the task list, it's always ensured that the task is visible if resizing
the terminal or when the task is moved due to a different status.