fix(core): kill child process correctly when run-script executor process is killed and not using pseudoterminal (#32699)
## Current Behavior

When running an npm script through nx and not using the PseudoTerminal,
child processes might not be killed when the `nx:run-script` executor is
killed.

## Expected Behavior

When running an npm script through nx and not using the PseudoTerminal,
child processes should be killed when the `nx:run-script` executor is
killed.

(cherry picked from commit 5d7ba935d668557f36c8a1e30164209b796574de)