feat(js)!: remove the experimental and deprecated inlining feature from the `tsc` and `swc` executors (#32783)
## Current Behavior

The experimental inlining feature in the `@nx/js:tsc` and `@nx/js:swc`
executors is deprecated.

## Expected Behavior

The experimental inlining feature in the `@nx/js:tsc` and `@nx/js:swc`
executors should be removed.

BREAKING CHANGE: The experimental inlining feature in the `@nx/js:tsc`
and `@nx/js:swc` executors was removed. A migration will remove the
related options (`external` and `externalBuildTargets`). Still, if you
use or rely on the feature, you need to make your dependencies buildable
or use a different build tool with bundling capabilities.