fix(core): ensure nx is published with the correct dependency version for the native packages (#32928)
## Current Behavior

The `nx` package is being published with the native package dependencies
using an invalid version range `"*"`.

## Expected Behavior

The `nx` package should be published with the native package
dependencies pointing to the same version as the `nx` package.

A recent change to `nx release` requires workspace packages to be
identified as such when deciding to replace the version.

## Related Issue(s)

Fixes #32898