chore(repo): add dotnet commit scope (#32975)
## Current Behavior
The PR to add the .NET plugin is failing because it uses the scope
`dotnet`, and that has to be present in master since the pr-title-checks
workflow validates against scopes in master.

## Expected Behavior
The scope is in master

## Related Issue(s)
<!-- Please link the issue being fixed so it gets closed when this is
merged. -->

Fixes #

(cherry picked from commit b7a291f8c17bcaa436623f26395aea6b587b79e9)