fix(core): display bottom corner indicator of tui in progress section correctly (#32985)
## Current Behavior

The bottom corner indicator of the in-progress section in the TUI is not
displayed when there are more tasks in progress than the maximum
parallel capacity (this can occur when there are continuous tasks).

## Expected Behavior

The bottom corner indicator of the in-progress section in the TUI should
always be displayed at the end of the running tasks.

(cherry picked from commit 8881e4233c60780fa5f6739ba89dd2d937e7b434)